+++ src/plone/rest/events.py
@@ -1,34 +1,41 @@
 # -*- coding: utf-8 -*-
 from plone.rest.interfaces import IAPIRequest
-from plone.rest.interfaces import IPUT
-from plone.rest.interfaces import IGET
-from plone.rest.interfaces import IPOST
 from plone.rest.interfaces import IDELETE
+from plone.rest.interfaces import IGET
 from plone.rest.interfaces import IOPTIONS
 from plone.rest.interfaces import IPATCH
+from plone.rest.interfaces import IPOST
+from plone.rest.interfaces import IPUT
+from plone.rest.negotiation import lookup_service_id
 from zope.interface import alsoProvides
 
 
 def mark_as_api_request(event):
-    """Mark a request with Accept 'application/json' with the IAPIRequest
-       interface.
+    """Mark a request as IAPIRequest if there's a service registered for the
+       actual request method and Accept header.
     """
     request = event.request
-    if request.getHeader('Accept') == 'application/json':
+    method = request.get('REQUEST_METHOD', 'GET')
+    accept = request.getHeader('Accept', 'text/html')
+    service_id = lookup_service_id(method, accept)
+
+    if service_id is not None:
         alsoProvides(request, IAPIRequest)
-        if request.get('REQUEST_METHOD') == 'PUT':
-            alsoProvides(request, IPUT)
-        if request.get('REQUEST_METHOD') == 'DELETE':
+        request._rest_service_id = service_id
+
+        if method == 'DELETE':
             alsoProvides(request, IDELETE)
-        if request.get('REQUEST_METHOD') == 'GET':
+        elif method == 'GET':
             alsoProvides(request, IGET)
-        if request.get('REQUEST_METHOD') == 'POST':
-            alsoProvides(request, IPOST)
-        if request.get('REQUEST_METHOD') == 'OPTIONS':
+        elif method == 'OPTIONS':
             alsoProvides(request, IOPTIONS)
-        if request.get('REQUEST_METHOD') == 'PATCH':
+        elif method == 'PATCH':
             alsoProvides(request, IPATCH)
+        elif method == 'POST':
+            alsoProvides(request, IPOST)
+        elif method == 'PUT':
+            alsoProvides(request, IPUT)
 
         # Flag as non-WebDAV request in order to avoid special treatment
         # in ZPublisher.BaseRequest.traverse().
-        event.request.maybe_webdav_client = 0
+        request.maybe_webdav_client = 0